hack/requirements.txt
absl-py==0.5.0
appdirs==1.4.3
appnope==0.1.0
asn1crypto==0.24.0
astor==0.7.1
attrs==18.1.0
beautifulsoup4==4.6.0
black==18.6b4
bleach==1.5.0
boto==2.49.0
boto3==1.9.17
botocore==1.12.17
bz2file==0.98
cachetools==2.1.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.1.4
cymem==1.31.2
cytoolz==0.9.0.1
dateparser==0.7.0
decorator==4.2.1
defusedxml==0.5.0
dill==0.2.8.2
Django==2.0
django-cors-headers==2.2.0
docutils==0.14
en-core-web-sm==2.0.0
entrypoints==0.2.3
flake8==3.5.0
gast==0.2.0
gensim==3.6.0
get==1.0.3
google-api-core==1.4.1
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
google-cloud-language==1.1.0
google-cloud-speech==0.36.0
google-images-download==2.3.0
googleapis-common-protos==1.5.3
grpcio==1.15.0
h5py==2.8.0
html5lib==0.9999999
httplib2==0.11.3
humanize==0.5.1
idna==2.7
ipykernel==4.8.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.2
jedi==0.11.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.1.6
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
lulu==0.5.3
Markdown==3.0.1
MarkupSafe==1.0
maya==0.5.0
mccabe==0.6.1
mistune==0.8.3
mnist==0.2.1
msgpack==0.5.6
msgpack-numpy==0.4.4.1
murmurhash==0.28.0
nbconvert==5.3.1
nbformat==4.4.0
nltk==3.3
notebook==5.4.0
numpy==1.15.2
oauth2client==4.1.3
oauthlib==2.1.0
opencv-contrib-python==3.4.1.15
pandocfilters==1.4.2
ParallelDots==3.2.11
parso==0.1.1
pbr==3.1.1
pendulum==1.5.1
pexpect==4.4.0
pickleshare==0.7.4
plac==0.9.6
post==1.0.2
preshed==1.0.1
prompt-toolkit==1.0.15
protobuf==3.6.1
ptyprocess==0.5.2
public==1.0.3
pyasn1==0.4.4
pyasn1-modules==0.2.2
PyAudio==0.2.11
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
Pygments==2.2.0
pygorithm==1.0.4
PyJWT==1.6.4
pyparsing==2.2.0
pyte==0.8.0
python-dateutil==2.7.3
python-dotenv==0.9.1
python3-openid==3.1.0
pytz==2018.5
pytzdata==2018.5
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
query-string==1.0.2
regex==2017.4.5
request==1.0.2
requests==2.19.1
requests-oauthlib==1.0.0
rsa==4.0
s3transfer==0.1.13
scipy==1.1.0
selenium==3.12.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
smart-open==1.7.1
snaptime==0.2.4
social-auth-app-django==2.1.0
social-auth-core==1.7.0
spacy==2.0.12
stevedore==1.28.0
summa==1.1.0
svgwrite==1.1.12
tensorboard==1.11.0
tensorflow==1.11.0
termcolor==1.1.0
terminado==0.8.1
termtosvg==0.4.0
testpath==0.3.1
thinc==6.10.3
toml==0.9.4
toolz==0.9.0
tornado==4.5.3
tqdm==4.26.0
traitlets==4.3.2
tzlocal==1.5.1
ujson==1.35
uritemplate==3.0.0
urllib3==1.23
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.8.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
wget==3.2
widgetsnbextension==3.1.4
wrapt==1.10.11
